\name{citccmst}
\docType{data}
\alias{citccmst}
\title{CIT colon expresson dataset used to classify data in Marisa et al 2012}
\description{
Colon Cancer expression dataset from CIT program used to define colon cancer molecular classification
}
\usage{citccmst}
\format{A list containing the expression data of the 443 samples and of the selected 57 probesets to classify data, the subtypes assigned to each sample and the Affymetrix annotation of the expression data from NetAffx (version na31)}
\source{CIT http://cit.ligue-cancer.net}
\references{
Marisa et al 2010 (submitted)
}
\keyword{datasets}


